requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
DateTime==4.5
dj-database-url==1.0.0
Django==4.0.5
django-apscheduler==0.6.2
Markdown==3.4.1
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
zope.interface==5.4.0
